requirements.txt
fastapi
uvicorn
requests
python-dotenv
jinja2